+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - config
 - save money popup
 - withdraw money popup
+- today spends component
 
 ## [3.1.0] - 2021-12-24
 ### Added